+++ src/app/app-routing.module.ts
@@ -1,10 +1,11 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { AppComponent } from './app.component';
+import { HomeComponent } from './home/home.component';
 import { KisahComponent } from './kisah/kisah.component';
 
 const routes: Routes = [
-  { path: 'home', component: AppComponent },
+  { path: '', component: HomeComponent },
+  { path: 'home', component: HomeComponent },
   { path: 'kisah', component: KisahComponent }
 ];
