+++ src/app/home/home.component.css
@@ -1,0 +1,9 @@
+/* Logo GKMI Anugerah */
+#logoAnugerah {
+    display: block;
+    max-width:230px;
+    max-height:95px;
+    width: auto;
+    height: auto;
+}
+
